+++ iqrs_wrangler.py
@@ -85,7 +85,9 @@ def lambda_handler(event, context):
 
         logger.info("Successfully sent data to sqs")
 
-        sqs.delete_message(QueueUrl=queue_url, ReceiptHandle=receipt_handler)
+        if receipt_handler:
+            sqs.delete_message(QueueUrl=queue_url, ReceiptHandle=receipt_handler)
+
         logger.info(funk.delete_data(bucket_name, in_file_name))
         logger.info("Successfully deleted input data from sqs")
         funk.send_sns_message(checkpoint, sns, arn)
